Renaming package names, artifactId and groupId
