+++ src/main/java/com/accesso/challengeladder/services/UserService.java
@@ -1,4 +1,4 @@
-package com.lopez83.spark.services;
+package com.accesso.challengeladder.services;
 
 import java.io.IOException;
 import java.sql.SQLException;
@@ -7,11 +7,11 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 
+import com.accesso.challengeladder.model.User;
+import com.accesso.challengeladder.utils.DBHelper;
 import com.j256.ormlite.dao.Dao;
 import com.j256.ormlite.dao.DaoManager;
 import com.j256.ormlite.support.ConnectionSource;
-import com.lopez83.spark.model.User;
-import com.lopez83.spark.utils.DBHelper;
 
 public class UserService {
