+++ src/main/java/com/accesso/challengeladder/utils/DBHelper.java
@@ -1,4 +1,4 @@
-package com.lopez83.spark.utils;
+package com.accesso.challengeladder.utils;
 
 import java.io.IOException;
 import java.sql.SQLException;
