+++ src/main/java/com/accesso/challengeladder/utils/JsonUtil.java
@@ -1,4 +1,4 @@
-package com.lopez83.spark.utils;
+package com.accesso.challengeladder.utils;
 
 import com.google.gson.Gson;
